+++ worlds/banjo_tooie/Options.py
@@ -10,11 +10,12 @@ class EnableMultiWorldJinjos(DefaultOnToggle):
     display_name = "Randomize Jinjos"
 
 class ForbidMovesOnJinjoFamilyTreasure(Choice):
-    """If Jinjos are randomized, do not allow unlockable moves behind Jinjo Families."""
-    display_name = "Forbid Unlockable Moves on Jinjo Family Treasure"
+    """If Jinjos are randomized, do not allow unlockable moves or magic behind Jinjo Families."""
+    display_name = "Forbid Unlockable Moves or Magic on Jinjo Family Treasure"
     option_none = 0
     option_moves_only = 1
     option_moves_and_magic = 2
+    option_magic_only = 3
     default = 1
 
 class ForbidJinjosOnJinjoFamilyTreasure(Toggle):
@@ -99,7 +100,7 @@ class BanjoTooieOptions(PerGameCommonOptions):
     death_link: DeathLink
     randomize_moves: EnableMultiWorldMoveList
     randomize_jinjos: EnableMultiWorldJinjos
-    forbid_moves_on_jinjo_family: ForbidMovesOnJinjoFamilyTreasure
+    forbid_on_jinjo_family: ForbidMovesOnJinjoFamilyTreasure
     forbid_jinjos_on_jinjo_family: ForbidJinjosOnJinjoFamilyTreasure
     randomize_doubloons: EnableMultiWorldDoubloons
     randomize_cheato: EnableMultiWorldCheatoPages